Expose insightDismissals on BackendProvider + CloudKitBackend (Phase D)
Adds the insightDismissals requirement to BackendProvider and wires it
through CloudKitBackend (protocol + concrete props, hooks, bundle field,
makeRepositories construction) and the ProfileSession hook fan-out. Fixes
the two BackendProvider test doubles (TestAuthBackend,
CloudKitAnalysisTestBackend) to forward the wrapped backend's repo, plus
the ProfileGRDBRepositories constructions in the sync-handler test support
and the two sync benchmarks that the Task 6 struct field left incomplete.

Extracts CloudKitBackend's repository-factory helpers into
CloudKitBackend+Repositories.swift to keep the class body under the
type_body_length limit after the new stored property and hooks.
