+++ Backends/CloudKit/CloudKitBackend.swift
@@ -5,6 +5,7 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
   let auth: any AuthProvider
   let accounts: any AccountRepository
   let accountGroups: any AccountGroupRepository
+  let insightDismissals: any InsightDismissalRepository
   let transactions: any TransactionRepository
   let categories: any CategoryRepository
   let transferSuggestions: any TransferSuggestionRepository
@@ -37,6 +38,7 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
   let grdbInstruments: GRDBInstrumentRegistryRepository
   let grdbAccounts: GRDBAccountRepository
   let grdbAccountGroups: GRDBAccountGroupRepository
+  let grdbInsightDismissals: GRDBInsightDismissalRepository
   let grdbCategories: GRDBCategoryRepository
   let grdbTransferSuggestions: GRDBTransferSuggestionRepository
   let grdbEarmarks: GRDBEarmarkRepository
@@ -60,6 +62,8 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
     let onAccountDeleted: @Sendable (String, UUID) -> Void
     let onAccountGroupChanged: @Sendable (String, UUID) -> Void
     let onAccountGroupDeleted: @Sendable (String, UUID) -> Void
+    let onInsightDismissalChanged: @Sendable (String, UUID) -> Void
+    let onInsightDismissalDeleted: @Sendable (String, UUID) -> Void
     let onCategoryChanged: @Sendable (String, UUID) -> Void
     let onCategoryDeleted: @Sendable (String, UUID) -> Void
     let onTransferSuggestionChanged: @Sendable (String, UUID) -> Void
@@ -84,6 +88,8 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
       onAccountDeleted: { _, _ in },
       onAccountGroupChanged: { _, _ in },
       onAccountGroupDeleted: { _, _ in },
+      onInsightDismissalChanged: { _, _ in },
+      onInsightDismissalDeleted: { _, _ in },
       onCategoryChanged: { _, _ in },
       onCategoryDeleted: { _, _ in },
       onTransferSuggestionChanged: { _, _ in },
@@ -119,6 +125,7 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
 
     self.grdbAccounts = repos.accounts
     self.grdbAccountGroups = repos.accountGroups
+    self.grdbInsightDismissals = repos.insightDismissals
     self.grdbTransactions = repos.transactions
     self.grdbCategories = repos.categories
     self.grdbTransferSuggestions = repos.transferSuggestions
@@ -132,6 +139,7 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
 
     self.accounts = repos.accounts
     self.accountGroups = repos.accountGroups
+    self.insightDismissals = repos.insightDismissals
     self.transactions = repos.transactions
     self.categories = repos.categories
     self.transferSuggestions = repos.transferSuggestions
@@ -147,149 +155,14 @@ final class CloudKitBackend: BackendProvider, @unchecked Sendable {
     self.groupUIState = GRDBGroupUIStateRepository(database: database)
   }
 
-  /// Bundle of GRDB repositories produced by `makeRepositories`. Keeps
-  /// the init body compact by handing back one value rather than ten.
-  private struct GRDBRepositoryBundle {
-    let accounts: GRDBAccountRepository
-    let accountGroups: GRDBAccountGroupRepository
-    let transactions: GRDBTransactionRepository
-    let categories: GRDBCategoryRepository
-    let transferSuggestions: GRDBTransferSuggestionRepository
-    let earmarks: GRDBEarmarkRepository
-    let earmarkBudgetItems: GRDBEarmarkBudgetItemRepository
-    let investments: GRDBInvestmentRepository
-    let transactionLegs: GRDBTransactionLegRepository
-    let analysis: GRDBAnalysisRepository
-    let insightDataSource: GRDBInsightDataSource
-    let csvImportProfiles: GRDBCSVImportProfileRepository
-    let importRules: GRDBImportRuleRepository
-  }
-
-  /// Constructs every GRDB-backed repository against the same writer
-  /// and hook fan-out, bundled so `init` only has to plumb the result
-  /// onto its stored properties.
-  private static func makeRepositories(
-    database: any DatabaseWriter,
-    instrument: Instrument,
-    conversionService: any InstrumentConversionService,
-    instrumentSeams: InstrumentSeams,
-    hooks: CloudKitBackendHooks
-  ) -> GRDBRepositoryBundle {
-    // The instrument-resolving read repos (accounts, transactions,
-    // earmarks, investments, analysis) all take the instrument seams;
-    // the remaining record-type repos don't, so the two groups are
-    // built by separate helpers.
-    let resolving = makeResolvingRepositories(
-      database: database,
-      instrument: instrument,
-      conversionService: conversionService,
-      instrumentSeams: instrumentSeams,
-      hooks: hooks)
-    return GRDBRepositoryBundle(
-      accounts: resolving.accounts,
-      accountGroups: GRDBAccountGroupRepository(
-        database: database,
-        onRecordChanged: hooks.onAccountGroupChanged,
-        onRecordDeleted: hooks.onAccountGroupDeleted),
-      transactions: resolving.transactions,
-      categories: GRDBCategoryRepository(
-        database: database,
-        onRecordChanged: hooks.onCategoryChanged,
-        onRecordDeleted: hooks.onCategoryDeleted),
-      transferSuggestions: GRDBTransferSuggestionRepository(
-        database: database,
-        onRecordChanged: hooks.onTransferSuggestionChanged,
-        onRecordDeleted: hooks.onTransferSuggestionDeleted),
-      earmarks: resolving.earmarks,
-      earmarkBudgetItems: GRDBEarmarkBudgetItemRepository(
-        database: database,
-        onRecordChanged: hooks.onEarmarkBudgetItemChanged,
-        onRecordDeleted: hooks.onEarmarkBudgetItemDeleted),
-      investments: resolving.investments,
-      transactionLegs: GRDBTransactionLegRepository(
-        database: database,
-        onRecordChanged: hooks.onTransactionLegChanged,
-        onRecordDeleted: hooks.onTransactionLegDeleted),
-      analysis: resolving.analysis,
-      insightDataSource: resolving.insightDataSource,
-      csvImportProfiles: GRDBCSVImportProfileRepository(
-        database: database,
-        onRecordChanged: hooks.onCSVImportProfileChanged,
-        onRecordDeleted: hooks.onCSVImportProfileDeleted),
-      importRules: GRDBImportRuleRepository(
-        database: database,
-        onRecordChanged: hooks.onImportRuleChanged,
-        onRecordDeleted: hooks.onImportRuleDeleted))
-  }
-
-  /// The five repositories that resolve instruments via the injected
-  /// `InstrumentMapResolving`. Grouped together semantically:
-  /// resolver-dependent repositories belong here, so a future repository
-  /// that needs the resolver is added here, not into the main
-  /// `makeRepositories` body.
-  private struct ResolvingRepositories {
-    let accounts: GRDBAccountRepository
-    let transactions: GRDBTransactionRepository
-    let earmarks: GRDBEarmarkRepository
-    let investments: GRDBInvestmentRepository
-    let analysis: GRDBAnalysisRepository
-    let insightDataSource: GRDBInsightDataSource
-  }
-
   /// The read-side instrument resolver and the write-side registrar,
   /// bundled to keep the repository-construction helpers' parameter
   /// lists small. In production both are the same shared
   /// `GRDBInstrumentRegistryRepository`; the type keeps them distinct
-  /// so the seams remain independently swappable.
+  /// so the seams remain independently swappable. The repository-factory
+  /// helpers that consume this live in `CloudKitBackend+Repositories`.
   struct InstrumentSeams {
     let resolver: any InstrumentMapResolving
     let registrar: any InstrumentRegistering
-  }
-
-  private static func makeResolvingRepositories(
-    database: any DatabaseWriter,
-    instrument: Instrument,
-    conversionService: any InstrumentConversionService,
-    instrumentSeams: InstrumentSeams,
-    hooks: CloudKitBackendHooks
-  ) -> ResolvingRepositories {
-    let resolver = instrumentSeams.resolver
-    return ResolvingRepositories(
-      accounts: GRDBAccountRepository(
-        database: database,
-        instrumentResolver: resolver,
-        instrumentRegistrar: instrumentSeams.registrar,
-        onRecordChanged: hooks.onAccountChanged,
-        onRecordDeleted: hooks.onAccountDeleted),
-      transactions: GRDBTransactionRepository(
-        database: database,
-        defaultInstrument: instrument,
-        conversionService: conversionService,
-        instrumentResolver: resolver,
-        instrumentRegistrar: instrumentSeams.registrar,
-        onRecordChanged: hooks.onTransactionChanged,
-        onRecordDeleted: hooks.onTransactionDeleted),
-      earmarks: GRDBEarmarkRepository(
-        database: database,
-        defaultInstrument: instrument,
-        instrumentResolver: resolver,
-        onRecordChanged: hooks.onEarmarkChanged,
-        onRecordDeleted: hooks.onEarmarkDeleted),
-      investments: GRDBInvestmentRepository(
-        database: database,
-        defaultInstrument: instrument,
-        instrumentResolver: resolver,
-        onRecordChanged: hooks.onInvestmentChanged,
-        onRecordDeleted: hooks.onInvestmentDeleted),
-      analysis: GRDBAnalysisRepository(
-        database: database,
-        instrument: instrument,
-        conversionService: conversionService,
-        instrumentResolver: resolver),
-      insightDataSource: GRDBInsightDataSource(
-        database: database,
-        instrument: instrument,
-        conversionService: conversionService,
-        instrumentResolver: resolver))
   }
 }
