+++ MoolahBenchmarks/SyncDownloadBenchmarks.swift
@@ -37,6 +37,7 @@ final class SyncDownloadBenchmarks: XCTestCase {
       categories: result.backend.grdbCategories,
       accounts: result.backend.grdbAccounts,
       accountGroups: result.backend.grdbAccountGroups,
+      insightDismissals: result.backend.grdbInsightDismissals,
       earmarks: result.backend.grdbEarmarks,
       earmarkBudgetItems: result.backend.grdbEarmarkBudgetItems,
       investmentValues: result.backend.grdbInvestments,
