+++ MoolahBenchmarks/SyncUploadBenchmarks.swift
@@ -38,6 +38,7 @@ final class SyncUploadBenchmarks: XCTestCase {
       categories: result.backend.grdbCategories,
       accounts: result.backend.grdbAccounts,
       accountGroups: result.backend.grdbAccountGroups,
+      insightDismissals: result.backend.grdbInsightDismissals,
       earmarks: result.backend.grdbEarmarks,
       earmarkBudgetItems: result.backend.grdbEarmarkBudgetItems,
       investmentValues: result.backend.grdbInvestments,
